# frontend/src/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "PolicyPilot API",
    "description" : "Policies, rule set versions, decisions and their traces. The model proposes and explains, the rules engine decides, a person approves every policy change.",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "This API"
  } ],
  "paths" : {
...
    "/api/v1/rulesets/{id}/versions/{no}/publish" : {
      "post" : {
        "tags" : [ "ruleset-controller" ],
        "summary" : "Publish a DRAFT version: compile, snapshot its rules and write the audit entry",
        "operationId" : "publish",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "no",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The published version",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VersionResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "No such rule set version in this sandbox",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409" : {
            "description" : "The version's status does not allow this",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422" : {
            "description" : "The rule set fails the Document 3 validator, or its review does not allow publishing yet",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/changes/{id}/approve" : {
      "post" : {
        "tags" : [ "change-controller" ],
        "summary" : "Approve a proposed change: publish the next version with its audit entry",
        "operationId" : "approve",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DecideChangeRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "The request, APPROVED, with the version the approval published",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ChangeDecisionResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "The note is too long or has a control character",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404" : {
            "description" : "No such change request in this sandbox",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409" : {
            "description" : "The request is not PROPOSED, or its base is no longer the latest version of its rule set",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/rulesets/{id}/versions/{no}" : {
      "get" : {
        "tags" : [ "ruleset-controller" ],
        "summary" : "A rule set version with its rules, findings and status",
        "operationId" : "version",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "no",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The version with its rules, findings and status",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VersionResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "No such rule set version in this sandbox",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/decisions/{id}" : {
      "get" : {
        "tags" : [ "decision-controller" ],
        "summary" : "A stored decision with its trace",
        "operationId" : "decision",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The decision object of Document 3",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            }
          },
          "404" : {
            "description" : "No such decision in this sandbox",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/changes/{id}/decisions/{decisionId}/trace" : {
      "get" : {
        "tags" : [ "change-controller" ],
        "summary" : "What a proposal decides for one of the sandbox's decisions on its base version",
        "operationId" : "proposedTrace",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "decisionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The decision object of Document 3 on the patched copy, with its trace, basedOnDecisionId and the request's number; nothing is stored",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            }
          },
          "404" : {
            "description" : "No such change request or decision in this sandbox, or a decision of another version",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "AcknowledgementResponse" : {
        "type" : "object",
        "properties" : {
          "resolution" : {
            "type" : "string",
            "enum" : [ "rule_added", "flag_added", "interpretation" ]
          },
          "note" : {
            "type" : "string"
          },
          "at" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "at" ]
      },
      "FindingResponse" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "severity" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "ruleIds" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "fieldNames" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        },
        "required" : [ "code", "fieldNames", "message", "path", "ruleIds", "severity" ]
      },
      "ReviewFindingResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "kind" : {
            "type" : "string",
            "enum" : [ "ambiguity", "conflict", "unsupported", "gap", "duplicate", "injection" ]
          },
          "severity" : {
            "type" : "string",
            "enum" : [ "error", "warning" ]
          },
          "ruleIds" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "paragraphIndexes" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int32"
            }
          },
          "message" : {
            "type" : "string"
          },
          "suggestion" : {
            "type" : "string"
          },
          "confidence" : {
            "type" : "number",
            "format" : "double"
          },
          "blocking" : {
            "type" : "boolean"
          },
          "acknowledgement" : {
            "$ref" : "#/components/schemas/AcknowledgementResponse"
          }
        },
        "required" : [ "blocking", "confidence", "id", "kind", "message", "paragraphIndexes", "ruleIds", "severity", "suggestion" ]
      },
      "ReviewResponse" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string",
            "enum" : [ "DONE", "FAILED", "STALE" ]
          },
          "promptVersion" : {
            "type" : "string"
          },
          "findings" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ReviewFindingResponse"
            }
          },
          "coverage" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          }
        },
        "required" : [ "coverage", "findings", "promptVersion", "status" ]
      },
      "VersionResponse" : {
        "type" : "object",
        "properties" : {
          "rulesetId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "domain" : {
            "type" : "string"
          },
          "protected" : {
            "type" : "boolean"
          },
          "forkedFromId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "versionId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "versionNo" : {
            "type" : "integer",
            "format" : "int32"
          },
          "status" : {
            "type" : "string"
          },
          "policyVersionId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "parentVersionId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "publishedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "publishedBy" : {
            "type" : "string"
          },
          "ruleSet" : {
            "description" : "The whole DSL document (Document 3)"
          },
          "findings" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/FindingResponse"
            }
          },
          "review" : {
            "$ref" : "#/components/schemas/ReviewResponse"
          }
        },
        "required" : [ "domain", "findings", "name", "policyVersionId", "protected", "ruleSet", "rulesetId", "status", "versionId", "versionNo" ]
      },
      "ErrorDetail" : {
        "type" : "object",
        "properties" : {
          "path" : {
            "type" : "string"
          },
          "problem" : {
            "type" : "string"
          }
        },
        "required" : [ "path", "problem" ]
      },
      "ErrorEnvelope" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "details" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ErrorDetail"
            }
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "required" : [ "code", "details", "message", "traceId" ]
      },
...
      "DecideChangeRequest" : {
        "type" : "object",
        "properties" : {
          "note" : {
            "type" : "string"
          }
        }
      },
      "ChangeDecisionResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "APPROVED", "REJECTED" ]
          },
          "decidedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "result" : {
            "$ref" : "#/components/schemas/Result"
          }
        },
        "required" : [ "decidedAt", "id", "number", "status" ]
      },
      "Result" : {
        "type" : "object",
        "properties" : {
          "rulesetId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "versionNo" : {
            "type" : "integer",
            "format" : "int32"
          },
          "versionId" : {
            "type" : "string",
            "format" : "uuid"
          }
        },
        "required" : [ "rulesetId", "versionId", "versionNo" ]
      },
...
    }
  }
}